# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API - Sistema de Gestão de Ordens de Serviço",
    "version": "1.0.0",
    "description": "API REST para gerenciamento de ordens de serviço de assistência técnica.\n\n**Autores:** Diogo Pinho, Yan Marra, Gabriel Vargas"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor principal"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "complement": {
            "type": "string",
            "nullable": true
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TechnicianRef": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "ServiceOrderHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "previousStatus": {
            "type": "string",
            "nullable": true
          },
          "newStatus": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "changedByUser": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              }
            }
          }
        }
      },
      "ServiceOrderPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "ServiceOrderPart": {
            "type": "object",
            "description": "Dados da associação (tabela de junção)",
            "properties": {
              "quantity": {
                "type": "integer"
              },
              "unitPrice": {
                "type": "number",
                "format": "float"
              }
            }
          }
        }
      },
      "ServiceOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string",
            "example": "OS000001"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "equipment": {
            "type": "string"
          },
          "problemDescription": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "concluded",
              "canceled"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ]
          },
          "technicianId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "openedAt": {
            "type": "string",
            "format": "date-time"
          },
          "concludedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelReason": {
            "type": "string",
            "nullable": true
          },
          "attendanceTimeMinutes": {
            "type": "integer",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "technician": {
            "$ref": "#/components/schemas/TechnicianRef"
          },
          "creator": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceOrderPart"
            }
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceOrderHistory"
            }
          }
        }
      },
      "ServiceOrderList": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 12
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceOrder"
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/service-orders": {
      "get": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Listar ordens de serviço",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "in_progress",
                "concluded",
                "canceled"
              ]
            }
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string",
              "enum": [
                "low",
                "medium",
                "high",
                "urgent"
              ]
            }
          },
          {
            "in": "query",
            "name": "technicianId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de ordens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceOrderList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Abrir nova ordem de serviço",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "clientId",
                  "equipment",
                  "problemDescription"
                ],
                "properties": {
                  "clientId": {
                    "type": "string"
                  },
                  "equipment": {
                    "type": "string",
                    "example": "Notebook Dell Inspiron"
                  },
                  "problemDescription": {
                    "type": "string",
                    "example": "Não liga ao pressionar o botão de energia"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high",
                      "urgent"
                    ],
                    "default": "medium"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ordem criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceOrder"
                }
              }
            }
          }
        }
      }
    },
    "/service-orders/{id}": {
      "get": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Buscar ordem por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da ordem com histórico e peças",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceOrder"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Atualizar ordem (bloqueada se concluída ou cancelada)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "equipment": {
                    "type": "string"
                  },
                  "problemDescription": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high",
                      "urgent"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ordem atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceOrder"
                }
              }
            }
          },
          "403": {
            "description": "Ordem concluída não pode ser editada"
          }
        }
      }
    },
    "/service-orders/{id}/assign": {
      "patch": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Atribuir técnico à ordem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "technicianId"
                ],
                "properties": {
                  "technicianId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Técnico atribuído com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceOrder"
                }
              }
            }
          },
          "422": {
            "description": "Técnico já possui 5 ordens em andamento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/service-orders/{id}/conclude": {
      "patch": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Concluir ordem de serviço",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordem concluída. Tempo de atendimento calculado automaticamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceOrder"
                }
              }
            }
          },
          "422": {
            "description": "Ordem sem técnico responsável não pode ser concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/service-orders/{id}/cancel": {
      "patch": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Cancelar ordem de serviço (motivo obrigatório)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "cancelReason"
                ],
                "properties": {
                  "cancelReason": {
                    "type": "string",
                    "example": "Cliente desistiu do reparo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ordem cancelada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceOrder"
                }
              }
            }
          },
          "422": {
            "description": "Motivo obrigatório",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/service-orders/{id}/history": {
      "get": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Histórico de alterações de status da ordem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de movimentações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceOrderHistory"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/service-orders/{id}/parts": {
      "post": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Vincular peça à ordem (desconta estoque)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "partId",
                  "quantity"
                ],
                "properties": {
                  "partId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Peça vinculada com sucesso"
          },
          "422": {
            "description": "Peça sem estoque ou estoque insuficiente"
          }
        }
      }
    },
...
  },
...
}